Line Tracker Modules
Added prototype 1 and 2 modules for error detection and line tracking.

lineTracker1.cpp First line tracking algorythm designed to calculate the priority of turn based off white multiplied by thier index off center of screen. White can be detected either in terms of Threshold or in terms of hard coded values. It was found that 100-150 was roughly the best range.

lineTracker2.cpp Second line tracking algoryhtm, this was writted using a PD based system which ranks priority of turn in terms of index of pixel. The total calculated error is expressed in terms of proportional error and change of (derivative) error. This will then be outputted to the motors. This will need to be tested.